2021.8.9

+++ Solutions/Count-Number-of-Special-Subsequences.cpp
@@ -2,7 +2,7 @@ class Solution {
 public:
     int countSpecialSubsequences(vector<int>& nums) {
         int64_t t0 = 0, t1 = 0, t2 = 0, t2_t1_t0 = 0;
-        int len = nums.size(), mod = 1000000000 + 7;
+        int64_t len = nums.size(), mod = 1000000000 + 7;
         for (auto num : nums)
         {
             if (num == 0) {
